combined C&S and Feed Storage data structs into same module
