+++ RUFAS/routines/feed_storage/grain.py
@@ -1,4 +1,4 @@
-from RUFAS.enums import CropCategory
+from RUFAS.data_structures.crop_soil_feed_storage_connection import CropCategory
 from .storage import Storage
