+++ RUFAS/routines/feed_storage/silage.py
@@ -1,13 +1,12 @@
 from typing import Optional
 
 from RUFAS.general_constants import GeneralConstants
-from RUFAS.data_structures.harvested_crop import HarvestedCrop
+from RUFAS.data_structures.crop_soil_feed_storage_connection import CropCategory, HarvestedCrop
 from RUFAS.output_manager import OutputManager
 from RUFAS.time import Time
 from RUFAS.units import MeasurementUnits
 from RUFAS.weather import Weather
 
-from RUFAS.enums import CropCategory
 from .storage import Storage
 
 """Fraction of effluent that is dry matter by mass."""
