+++ RUFAS/routines/feed_storage/storage.py
@@ -1,14 +1,13 @@
 from typing import List
 
 from RUFAS.current_day_conditions import CurrentDayConditions
-from RUFAS.data_structures.harvested_crop import HarvestedCrop
+from RUFAS.data_structures.crop_soil_feed_storage_connection import CropCategory, CropType, HarvestedCrop
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.output_manager import OutputManager
 from RUFAS.time import Time
 from RUFAS.units import MeasurementUnits
 from RUFAS.weather import Weather
 
-from RUFAS.enums import CropCategory, CropType
 
 """
 These constants define the upper and lower bounds of temperatures that allow fermentation (in degrees C), the upper and
