+++ RUFAS/routines/field/crop/crop.py
@@ -4,8 +4,7 @@ from copy import copy
 from typing import Any, Optional
 
 from RUFAS.current_day_conditions import CurrentDayConditions
-from RUFAS.data_structures.harvested_crop import HarvestedCrop
-from RUFAS.enums import StorageType
+from RUFAS.data_structures.crop_soil_feed_storage_connection import HarvCropStorageType
 from RUFAS.routines.field.crop.biomass_allocation import BiomassAllocation
 from RUFAS.routines.field.crop.crop_data import CropData
 from RUFAS.routines.field.crop.crop_enum import CropSpecies
@@ -206,7 +205,7 @@ class Crop:
         field_size: float,
         time: Time,
         soil_data: SoilData,
-    ) -> tuple[HarvestedCrop, StorageType] | None:
+    ) -> HarvCropStorageType | None:
         """Wrapper function for the Crop's CropManagement harvesting operation.
 
         Parameters
