+++ RUFAS/routines/field/manager/field_manager.py
@@ -1,7 +1,6 @@
 from typing import Dict, List, Tuple
 
-from RUFAS.enums import StorageType
-from RUFAS.data_structures.harvested_crop import HarvestedCrop
+from RUFAS.data_structures.crop_soil_feed_storage_connection import HarvCropStorageType
 from RUFAS.input_manager import InputManager
 from RUFAS.output_manager import OutputManager
 from RUFAS.routines.field.field.field import Field
@@ -57,7 +56,7 @@ class FieldManager:
             self.fields.append(new_field)
         self.output_gatherer = FieldDataReporter(fields=self.fields)
 
-    def daily_update_routine(self, weather: Weather, time: Time) -> list[tuple[HarvestedCrop, StorageType]]:
+    def daily_update_routine(self, weather: Weather, time: Time) -> list[HarvCropStorageType]:
         """
         This method will run the daily routine in the field, which will be calling the manage field method on each
         field.
@@ -74,7 +73,7 @@ class FieldManager:
         Because different fields can have different latitudes, the day length has to be recalculated for each field.
 
         """
-        harvested_crops: list[tuple[HarvestedCrop, StorageType]] = []
+        harvested_crops: list[HarvCropStorageType] = []
         for field in self.fields:
             current_conditions = weather.get_current_day_conditions(time, field.field_data.absolute_latitude)
             info_map = {
